Complete Rock Paper Scissors Logic in JS file
I am so mad lol I forgot to keep a consistent string of small commits
and accidentally finished the whole thing! I need to work on commiting
more often and using more detail because this bad habit is not that big
of a deal now but will bite me in the butt when I get to more complicated
projects later.